[Obs AI Assistant] Extract `get_data_on_screen` function (#214362)
Minor refactor to extract get registration of `get_data_on_screen`. This
way the function co-located with the other LLM functions making it
easier to find and test.

(cherry picked from commit 6bb4badddcf31fe0a133fab8d580250981f2dcb9)
