+++ x-pack/platform/plugins/shared/observability_ai_assistant/server/functions/index.ts
@@ -13,7 +13,7 @@ import { registerElasticsearchFunction } from './elasticsearch';
 import { GET_DATASET_INFO_FUNCTION_NAME, registerGetDatasetInfoFunction } from './get_dataset_info';
 import { registerKibanaFunction } from './kibana';
 import { registerExecuteConnectorFunction } from './execute_connector';
-import { GET_DATA_ON_SCREEN_FUNCTION_NAME } from '../service/chat_function_client';
+import { GET_DATA_ON_SCREEN_FUNCTION_NAME } from './get_data_on_screen';
 
 // cannot be imported from x-pack/solutions/observability/plugins/observability_ai_assistant_app/server/functions/query/index.ts due to circular dependency
 export const QUERY_FUNCTION_NAME = 'query';
