+++ x-pack/platform/plugins/shared/observability_ai_assistant/server/service/chat_function_client/index.ts
@@ -7,10 +7,9 @@
 /* eslint-disable max-classes-per-file*/
 
 import Ajv, { type ErrorObject, type ValidateFunction } from 'ajv';
-import dedent from 'dedent';
 import { compact, keyBy } from 'lodash';
 import { Logger } from '@kbn/logging';
-import { FunctionVisibility, type FunctionResponse } from '../../../common/functions/types';
+import { type FunctionResponse } from '../../../common/functions/types';
 import type { Message, ObservabilityAIAssistantScreenContextRequest } from '../../../common/types';
 import { filterFunctionDefinitions } from '../../../common/utils/filter_function_definitions';
 import type {
@@ -21,6 +20,7 @@ import type {
   RegisterFunction,
   RegisterInstruction,
 } from '../types';
+import { registerGetDataOnScreenFunction } from '../../functions/get_data_on_screen';
 
 export class FunctionArgsValidationError extends Error {
   constructor(public readonly errors: ErrorObject[]) {
@@ -32,8 +32,6 @@ const ajv = new Ajv({
   strict: false,
 });
 
-export const GET_DATA_ON_SCREEN_FUNCTION_NAME = 'get_data_on_screen';
-
 export class ChatFunctionClient {
   private readonly instructions: InstructionOrCallback[] = [];
 
@@ -43,47 +41,9 @@ export class ChatFunctionClient {
   private readonly actions: Required<ObservabilityAIAssistantScreenContextRequest>['actions'];
 
   constructor(private readonly screenContexts: ObservabilityAIAssistantScreenContextRequest[]) {
-    const allData = compact(screenContexts.flatMap((context) => context.data));
-
     this.actions = compact(screenContexts.flatMap((context) => context.actions));
 
-    if (allData.length) {
-      this.registerFunction(
-        {
-          name: GET_DATA_ON_SCREEN_FUNCTION_NAME,
-          description: `Retrieve the structured data of content currently visible on the user's screen. Use this tool to understand what the user is viewing at this moment to provide more accurate and context-aware responses to their questions.`,
-          visibility: FunctionVisibility.AssistantOnly,
-          parameters: {
-            type: 'object',
-            properties: {
-              data: {
-                type: 'array',
-                description:
-                  'The pieces of data you want to look at it. You can request one, or multiple',
-                items: {
-                  type: 'string',
-                  enum: allData.map((data) => data.name),
-                },
-              },
-            },
-            required: ['data' as const],
-          },
-        },
-        async ({ arguments: { data: dataNames } }) => {
-          return {
-            content: allData.filter((data) => dataNames.includes(data.name)),
-          };
-        }
-      );
-
-      this.registerInstruction(({ availableFunctionNames }) =>
-        availableFunctionNames.includes(GET_DATA_ON_SCREEN_FUNCTION_NAME)
-          ? `The ${GET_DATA_ON_SCREEN_FUNCTION_NAME} function will retrieve specific content from the user's screen by specifying a data key. Use this tool to provide context-aware responses. Available data: ${dedent(
-              allData.map((data) => `${data.name}: ${data.description}`).join('\n')
-            )}`
-          : undefined
-      );
-    }
+    registerGetDataOnScreenFunction(this, screenContexts);
 
     this.actions.forEach((action) => {
       if (action.parameters) {
